Feature: add trait `RaftPayload` `RaftEntry` to access payload and entry without the need to know about user data, i.e., `AppData` or `AppDataResponse`.
- With `RaftPayload` and `RaftEntry`, the protocol can be implememnted
  in a standalone mod, without depending on type of user data.
  I.e. the protocol mod does not need to know about `AppData`, or
  `RaftStorage<RaftTypeConfig>`.

  This will make test easier to write.

- Refactor: move `Entry`, `EntryPayload`, `RaftPayload`, `RaftEntry`
  into file `entry.rs`.
