+++ openraft/src/core/admin.rs
@@ -23,10 +23,10 @@ use crate::metrics::RemoveTarget;
 use crate::raft::AddLearnerResponse;
 use crate::raft::ChangeMembers;
 use crate::raft::ClientWriteResponse;
-use crate::raft::EntryPayload;
 use crate::raft::RaftRespTx;
 use crate::raft_types::LogIdOptionExt;
 use crate::versioned::Updatable;
+use crate::EntryPayload;
 use crate::LogId;
 use crate::Membership;
 use crate::Node;
