+++ openraft/src/core/append_entries.rs
@@ -4,10 +4,10 @@ use crate::core::State;
 use crate::error::AppendEntriesError;
 use crate::raft::AppendEntriesRequest;
 use crate::raft::AppendEntriesResponse;
-use crate::raft::Entry;
-use crate::raft::EntryPayload;
 use crate::raft_types::LogIdOptionExt;
 use crate::EffectiveMembership;
+use crate::Entry;
+use crate::EntryPayload;
 use crate::LogId;
 use crate::MessageSummary;
 use crate::RaftNetworkFactory;
