+++ openraft/src/core/client.rs
@@ -20,10 +20,10 @@ use crate::raft::AppendEntriesRequest;
 use crate::raft::AppendEntriesResponse;
 use crate::raft::ClientWriteRequest;
 use crate::raft::ClientWriteResponse;
-use crate::raft::Entry;
-use crate::raft::EntryPayload;
 use crate::raft::RaftRespTx;
 use crate::replication::RaftEvent;
+use crate::Entry;
+use crate::EntryPayload;
 use crate::MessageSummary;
 use crate::RPCTypes;
 use crate::RaftNetwork;
