+++ openraft/src/store_ext.rs
@@ -7,7 +7,6 @@ use std::sync::Arc;
 
 use crate::async_trait::async_trait;
 use crate::defensive::DefensiveCheckBase;
-use crate::raft::Entry;
 use crate::storage::LogState;
 use crate::storage::RaftLogReader;
 use crate::storage::RaftSnapshotBuilder;
@@ -15,6 +14,7 @@ use crate::storage::Snapshot;
 use crate::summary::MessageSummary;
 use crate::DefensiveCheck;
 use crate::EffectiveMembership;
+use crate::Entry;
 use crate::LogId;
 use crate::RaftStorage;
 use crate::RaftStorageDebug;
