+++ openraft/tests/append_entries/t10_conflict_with_empty_entries.rs
@@ -3,9 +3,9 @@ use std::sync::Arc;
 use anyhow::Result;
 use memstore::ClientRequest;
 use openraft::raft::AppendEntriesRequest;
-use openraft::raft::Entry;
-use openraft::raft::EntryPayload;
 use openraft::Config;
+use openraft::Entry;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::RaftNetwork;
