+++ openraft/tests/append_entries/t20_append_conflicts.rs
@@ -4,8 +4,8 @@ use std::time::Duration;
 use anyhow::Result;
 use maplit::btreeset;
 use openraft::raft::AppendEntriesRequest;
-use openraft::raft::Entry;
 use openraft::Config;
+use openraft::Entry;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::MessageSummary;
