+++ openraft/tests/append_entries/t30_append_inconsistent_log.rs
@@ -3,9 +3,9 @@ use std::time::Duration;
 
 use anyhow::Result;
 use maplit::btreeset;
-use openraft::raft::Entry;
-use openraft::raft::EntryPayload;
 use openraft::Config;
+use openraft::Entry;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::RaftLogReader;
