+++ openraft/tests/elect_compare_last_log.rs
@@ -5,9 +5,9 @@ use std::time::Duration;
 use anyhow::Result;
 use fixtures::RaftRouter;
 use maplit::btreeset;
-use openraft::raft::Entry;
-use openraft::raft::EntryPayload;
 use openraft::Config;
+use openraft::Entry;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::Membership;
