+++ openraft/tests/fixtures/mod.rs
@@ -39,8 +39,6 @@ use openraft::raft::AddLearnerResponse;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::AppendEntriesResponse;
 use openraft::raft::ClientWriteRequest;
-use openraft::raft::Entry;
-use openraft::raft::EntryPayload;
 use openraft::raft::InstallSnapshotRequest;
 use openraft::raft::InstallSnapshotResponse;
 use openraft::raft::VoteRequest;
@@ -49,6 +47,8 @@ use openraft::storage::RaftLogReader;
 use openraft::storage::RaftStorage;
 use openraft::Config;
 use openraft::DefensiveCheckBase;
+use openraft::Entry;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::LogIdOptionExt;
