+++ openraft/tests/initialization.rs
@@ -5,9 +5,9 @@ use std::time::Duration;
 use anyhow::Result;
 use fixtures::RaftRouter;
 use maplit::btreeset;
-use openraft::raft::EntryPayload;
 use openraft::Config;
 use openraft::EffectiveMembership;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::Membership;
