+++ openraft/tests/snapshot/snapshot_overrides_membership.rs
@@ -5,9 +5,9 @@ use std::time::Duration;
 use anyhow::Result;
 use maplit::btreeset;
 use openraft::raft::AppendEntriesRequest;
-use openraft::raft::Entry;
-use openraft::raft::EntryPayload;
 use openraft::Config;
+use openraft::Entry;
+use openraft::EntryPayload;
 use openraft::LeaderId;
 use openraft::LogId;
 use openraft::Membership;
